Create Message Central Child Variable
Made the three random message groups variable lengths.  You can select the number of variables you want for each group independently.  There are default value for the first 25 items on Group1 and Group2 and 10 for Group 3.  Any of these can be changed and those changes are retained.  Theoretically you can have an unlimited number of items in each group.  Generic default values will be populated for the number of items you select.