Update Weather-Display With External Forecast
[UPDATED]  V3.0.4   03/24/2019

Further reduction on the character limit for myTile.  This was primarily though renaming the icons to very short names.  Because of this you will need to re-copy the icons from my GitHub repository  (https://github.com/Scottma61/WeatherIcons) to your webserver to get the new named icons.

If the MyTile exceeds 1,024 characters (including file paths to icons and HTML formatting) then it attempt to display with no icons and no formatting.  It it is still too long, it will truncate it.